DOC: change animation writer from imagemagick to pillow
Due to a change from matplotlib 3.1.x to 3.2.0.
imagemagick was not installed in the continuous integration
environment previously, so FuncAnimation.save defaulted to
writer='pillow'. However, with matplotlib 3.2.0, this does not happen,
giving an error:
"TypeError: 'MovieWriterRegistry' object is not an iterator".

So the writer had to be set explicitly to 'pillow'.