Major rework of the Gaussian fitting refinement to improve performance
Previously the full image data was cropped to just the relevant parts
quite far into the fitting functionality, which lead to the refinement
part processing times scaling with the size of the image. So for large
images, this drastically reduced performance and lead to very long
processing times for large images (over 1000 x 1000 pixels).

This commit moves the cropping to much earlier in the process, which
drastically improve the performance. Albeit at a cost of slightly more
complicated code, as more variables needs to be kept track of.
Now, the processing scales much less with the size of the image, but
rather as a function of the number of atoms to be fitted.

Processing time comparisons:

500 x 500 image, 1089 atoms: 9 s (old), 7 s (new)
1000 x 1000 image, 4356 atoms: 72 s (old), 27 s (new)
2000 x 2000 image, 17424 atoms: 875 s (old), 108 s (new)